requirements.txt
redis==4.1.0
django==3.2.11
django_celery_beat==2.2.1
celery==5.1.2
django_celery_results==2.2.0
django_crispy_forms==1.13.0
django_bootstrap_datepicker_plus==4.0.0
django_bootstrap4==21.1
django_environ==0.8.1